backend/requirements.txt
# Phase 0 — Core server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.0
httpx==0.27.0
pillow==10.4.0
pdf2image==1.17.0

# Phase 1 — OCR (Aryan)
# google-cloud-vision==3.7.4
# numpy==1.26.4

# Phase 2 & 3 — Classification + Extraction (Ayan)
# transformers==4.44.0
# torch==2.4.0
# sentencepiece==0.2.0
# datasets==2.20.0

# Phase 5 — Database
# sqlalchemy==2.0.31
# alembic==1.13.2
# psycopg2-binary==2.9.9
